FEAT/Project - Implmenting ProjectRegistrationService. Also  implementing fluent build parttern to build and validate  ProjectRegistrationEntity.

+++ api/src/main/java/com/americas/challenge/api/model/dto/projectRegisterWorkRequestDTO.java
@@ -13,6 +13,6 @@ public class projectRegisterWorkRequestDTO {
     
     private Integer projectId;
     private Date date;
-    private Double time;
+    private Float time;
 
 }
